added new markers to app JS, cr and changed the way dispatchOrder caching works

+++ src/main/java/com/amt/entities/buisiness/Vehicle.java
@@ -16,8 +16,8 @@ public class Vehicle extends BasicEntity{
     public String vehicleLicensePlate;
     public long driver;
     public long status;
-    public long latitude;
-    public long longitude;
+    public double latitude;
+    public double longitude;
